Optimizied posture slider drag
